Complete stack test
