+++ exercisesSubmission.json
@@ -33,7 +33,7 @@
             "completed": true
         },
         "stack-test": {
-            "completed": false
+            "completed": true
         },
         "sum-test": {
             "completed": false
