+++ stack-test/stackTest.js
@@ -1,8 +1,33 @@
+class Stack {
+  constructor() {
+    this.items = [];
+    this.count = 0;
+  }
 
-function stackTest(...args) {
-  return args;
+  getLength() {
+    return this.count;
+  }
+
+  push(item) {
+    this.items.push(item);
+    this.count = this.count + 1;
+  }
+
+  pop() {
+    if (this.count > 0) {
+      this.count = this.count - 1;
+    }
+
+    return this.items.pop();
+  }
+
+  // returns top element in the stack
+  peek() {
+    return this.items.slice(-1)[0];
+  }
 }
 
+
 export {
-  stackTest,
+  Stack,
 };
